+++ lib/pages/settings/data/settings_controller.dart
@@ -1,8 +1,7 @@
 import 'package:chat_interface/pages/settings/account/data_settings.dart';
+import 'package:chat_interface/pages/settings/app/general_settings.dart';
 import 'package:chat_interface/pages/settings/town/file_settings.dart';
-import 'package:chat_interface/pages/settings/app/language_settings.dart';
 import 'package:chat_interface/pages/settings/app/log_settings.dart';
-import 'package:chat_interface/pages/settings/town/spaces_settings.dart';
 import 'package:chat_interface/pages/settings/app/speech_settings.dart';
 import 'package:chat_interface/pages/settings/town/tabletop_settings.dart';
 import 'package:chat_interface/pages/settings/app/video_settings.dart';
@@ -36,8 +35,7 @@ class SettingController extends GetxController {
     AudioSettings.addSettings(this);
     addVideoSettings(this);
     addCallAppearanceSettings(this);
-    addLanguageSettings(this);
-    SpacesSettings.addSpacesSettings(this);
+    GeneralSettings.addSettings(this);
     TabletopSettings.addSettings(this);
     ThemeSettings.addThemeSettings(this);
     FileSettings.addSettings(this);